Move fetch_and_parse
